refactor: eliminate utilities/ subdirectories
Problem
-------
In terms of project layout, it is unnecessary to have a utilities/
subdirectory (or "module family") in various sub-topics.
Utils should live in one of:
1. the top-level (root) util/ directory, or...
2. a named module ("foo.ts", no need for "util/foo.ts"), or...
3. "util.ts" in the family subdir (not "util/*.ts")

Solution
--------
- Move lambda/utilities/parameterUtils.ts -> lambda/config/parameterUtils.ts
- Move src/test/cdk/utilities/*.ts -> src/test/cdk/*.ts

This avoids unnecessary nesting and entropy.
